requirements.txt
numpy==1.14.3
PyMySQL==0.9.2
Django==2.1.7
djangorestframework==3.9.2
django-cors-headers
djangorestframework-jwt